move backend logic to a subpackage
Goal is to have one entry point for the backend state
and backend-specific implementations and keep the higher level
dedicated to the functional interface
