+++ fp/tensors/__init__.py
@@ -2,7 +2,7 @@ from .tensor import Tensor, Numpy, Jax, Torch, TorchBackend
 from .shape import Torus
 from .shape import Torus as Shape
 from .tens import Tens, Linear, Otimes
-from .backend import backend
+from .backend import backend, has_jax, has_torch, np, jnp, jax, torch 
 
 __all__ = [
    "Tensor",
