+++ fp/tensors/backend/__init__.py
@@ -1,0 +1,3 @@
+from ._pkgs import has_jax, has_torch, np, jnp, jax, torch
+from .interfaces import backend
+backend.init()
